staterecovery: refactor integration test
